Allow Tokeninput settings to be specified by TokenWidget.
Any kwarg (other than attrs) passed to TokenWidget.__init__ is assumed
to be a Tokeninput setting, and is passed along to the client-side via
a JSON-encoded data- attr. The keys are not checked for validity, to
ensure future compatibility with new Tokeninput features.

Settings can be specified using jQuery-style camelCase keys, or Python-
style underscored_keys. I like to use underscores.

A list of available settings can be found here:
http://loopj.com/jquery-tokeninput/