Spawn instances of the model for unrecognized keys.
This allows models (for example, tags) to be spawned and related within
the form itself. This requires JS support on the client-side, which
loopj/jquery-tokeninput@1273e689a0101067fa6db384f0d9108855678643
doesn't include. Check out loopj/jquery-tokeninput#219 for a patch.

If and when the patch is merged into master, I'll merge this branch.